Refactored code for aggregate error handling
